requirements.txt
mutagen==1.47.0
numba==0.60.0
numpy==2.2.1
Pillow==11.1.0
pycryptodome==3.20.0
pytest==8.3.4
Requests==2.32.3
setuptools==72.1.0
tqdm==4.66.4
